Stage 3 of unknown uploading all current files as of October 15th 2021
These files are only the up to date files as of October 15 2021. Project has been maintained elsewhere since roughly December of 2020 when 'V2' was first created. 'V2' was an updated version of 'V1' based on the original image carousel idea that I created in Godot, and development was shortly thereafter stopped for approx 6 months when I attempted to develop the idea in xcode. Current state of the app now needs to be processed in the github project, and all undeveloped features added to 'todo' before moving anything to 'in progress'. Current state needs to include mvp features, current issues/bugs, and anything else that needs doing before moving forward with development.